% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium.R
\name{calcBoundProtein}
\alias{calcBoundProtein}
\title{Calculate Bound Concentrations}
\usage{
calcBoundProtein(free_RNA, protein_totals, Kds)
}
\arguments{
\item{free_RNA}{Equilibrium free RNA concentration.}

\item{protein_totals}{Vector of total protein concentrations.}

\item{Kds}{Vector of dissociation constants.}
}
\value{
Vector of bound concentrations for each RBP.
}
\description{
Calculates the equilibrium bound concentrations for each RBP.
}
\examples{
free_rna <- solveEquilibrium(10, c(100, 100), c(10, 50))
calcBoundProtein(free_rna, c(100, 100), c(10, 50))
}
