% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countKmers}
\alias{countKmers}
\title{Count Unique K-mers}
\usage{
countKmers(kmers)
}
\arguments{
\item{kmers}{Character vector of k-mers.}
}
\value{
data.table with columns 'motif' and 'count'.
}
\description{
Returns a data.table with unique k-mers and their counts.
}
