% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractKmers}
\alias{extractKmers}
\title{Extract K-mers from Sequence}
\usage{
extractKmers(sequence, k = 5)
}
\arguments{
\item{sequence}{Character string of the RNA sequence.}

\item{k}{Integer size of the k-mer (default 5).}
}
\value{
Character vector of k-mers.
}
\description{
Generates a vector of all k-mers from a given sequence using a sliding window.
}
\examples{
extractKmers("ACGUACGU", k = 3)
}
