% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{featureScale}
\alias{featureScale}
\title{Feature Scaling}
\usage{
featureScale(X, MAX, MIN)
}
\arguments{
\item{X}{Numeric vector to scale.}

\item{MAX}{Maximum value of the new range.}

\item{MIN}{Minimum value of the new range.}
}
\value{
Scaled numeric vector.
}
\description{
Scales a numeric vector to a given range [MIN, MAX].
}
