% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateRNA}
\alias{generateRNA}
\title{Generate Random RNA Sequence}
\usage{
generateRNA(L, A = 0.25, G = 0.25, C = 0.25, U = 0.25)
}
\arguments{
\item{L}{Length of the sequence.}

\item{A}{Frequency of Adenine.}

\item{G}{Frequency of Guanine.}

\item{C}{Frequency of Cytosine.}

\item{U}{Frequency of Uracil.}
}
\value{
A character string representing the RNA sequence.
}
\description{
Generates a random RNA sequence with specified nucleotide frequencies.
}
