% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium.R
\name{solveEquilibrium}
\alias{solveEquilibrium}
\title{Equilibrium Binding Solver}
\usage{
solveEquilibrium(RNA_total, protein_totals, Kds)
}
\arguments{
\item{RNA_total}{Total RNA concentration (numeric).}

\item{protein_totals}{Vector of total protein concentrations for N RBPs (numeric vector).}

\item{Kds}{Vector of dissociation constants (Kd) for N RBPs (numeric vector).}
}
\value{
The equilibrium free RNA concentration.
}
\description{
Solves for the free RNA concentration in a competitive binding system.
}
\examples{
# Two RBPs competing for RNA
free_rna <- solveEquilibrium(
  RNA_total = 10,
  protein_totals = c(100, 100),
  Kds = c(10, 50)
)
free_rna
}
