% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Binding Heatmap}
\usage{
plotHeatmap(
  results,
  transcript,
  rbps = NULL,
  xlim = NULL,
  xaxis_type = c("nucleotide", "position", "both"),
  metric = c("occupancy", "density", "density_fc", "occupancy_fc"),
  window = NULL,
  zlim = NULL
)
}
\arguments{
\item{results}{data.table of simulation results.}

\item{transcript}{Transcript name to filter.}

\item{rbps}{Optional vector of RBP names.}

\item{xlim}{Optional x-axis limits.}

\item{xaxis_type}{Type of x-axis labels: "nucleotide", "position", or "both".}

\item{metric}{Type of metric: "occupancy", "density", "density_fc", or "occupancy_fc".}

\item{window}{Optional integer for K-mer sliding window average.}

\item{zlim}{Optional numeric vector for color scale limits.}
}
\value{
A ggplot object.
}
\description{
Plots a heatmap of binding for multiple RBPs.
}
\examples{
if (FALSE) {
model_file <- system.file("extdata", "model_RBP.csv", package = "RBPBind")
rbp_models <- setModel(loadModel(model_file))
results <- simulateBinding("ACGUACGUACGU", rbp_models, c(HH = 100, HL = 100))
plotHeatmap(results, transcript = NULL)
}
}
